Added CreateTransactionScreen

+++ core/model/src/main/java/com/example/model/BankAccountHistoryResponse.kt
@@ -1,9 +1,11 @@
 package com.example.model
 
+import java.math.BigDecimal
+
 data class BankAccountHistoryResponse(
     val accountId: Int,
     val accountName: String,
     val currency: String,
-    val currentBalance: Float,
+    val currentBalance: BigDecimal,
     val history: List<BankAccountHistory>
 )
