+++ core/model/src/main/java/com/example/model/StatItem.kt
@@ -1,10 +1,12 @@
 package com.example.model
 
+import java.math.BigDecimal
+
 data class StatItem(
     val categoryId: Int,
     val categoryName: String,
     val emoji: String,
-    val amount: Float
+    val amount: BigDecimal
 )
 
 fun StatItem.toCategory(): Category {
